refactor: 경험 추가 분석 및 저장 액션 분리
